refactor transactions route to typescript, adjust imports to use absolute paths
